requirements/development.txt
# Local development dependencies go here
-r base.txt
Sphinx==1.7.5
django-debug-toolbar==1.9.1
docutils==0.14
pygraphviz==1.3.1
pyyaml==3.12
